added tests for the common card, article and hero components
